Reformat and Consolidated Models
