+++ ArkansasAssetBuilders/src/main/java/model/TaxYear.java
@@ -1,49 +1,15 @@
 package model;
-
 import javafx.beans.property.*;
-
-/**
- * TaxYear class
- * <br/>
- * <br/>
- * Tax year data from inputted tax information.
- * Tax year information is stored in the local database.
- */
 public class TaxYear {
-    private IntegerProperty TaxYear;
+	private IntegerProperty tax_year;
+	public TaxYear(){
+		this.tax_year = new SimpleIntegerProperty();
+	}
+	public int getTaxYear(){return tax_year.get();}
 
-    /**
-     * Constructor for the TaxYear class.
-     */
-    public TaxYear(){
-        this.TaxYear = new SimpleIntegerProperty();
-    }
+	public void setTaxYear(int tax_year){this.tax_year.set(tax_year);}
 
-    /**
-     * Method for getting the tax year ID of the TaxYear object.
-     * @return int, tax year ID of the TaxYear object.
-     */
-    public int getTaxYear(){
-        return TaxYear.get();
-    }
-
-    /**
-     * Method for setting the tax year ID of the TaxYear object.
-     * @param TaxYear int, the tax year ID that the old tax year
-     *                  ID will be set to.
-     */
-    public void setTaxYear(int TaxYear){
-        this.TaxYear.set(TaxYear);
-    }
-
-    /**
-     * Method for retrieving TaxYear tax year ID from the database in an
-     * IntegerProperty format.<br/>
-     * @return IntegerProperty, the tax year ID of the TaxYear.
-     */
-    public IntegerProperty taxYearProperty(){
-        return TaxYear;
-    }
+	public IntegerProperty TaxYearProperty(){return tax_year;}
 
 
 }
